Rename tpt review licence
